+++ idearly/src/pages/AlgorithmSolvingPage/AlgorithmSolvingPage.tsx
@@ -1,10 +1,14 @@
-import { AlgorithmNav } from "./components";
+import { AlgorithmNav, AlgorithmTextChatModal } from "./components";
 import * as S from "./AlgorithmSolvingPage.styles";
+import { useDisclosure } from "@chakra-ui/react";
 
 export const AlgorithmSolvingPage = () => {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  
   return (
     <S.AlgorithmSolvingPageContainer>
-      <AlgorithmNav />
+      <AlgorithmTextChatModal isOpen={isOpen} onClose={onClose} />
+      <AlgorithmNav onOpen={onOpen} />
       <div>hi</div>
     </S.AlgorithmSolvingPageContainer>
   );
