+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmNav/AlgorithmNav.types.ts
@@ -1,0 +1,3 @@
+export interface Prop {
+  onOpen: () => void
+}
