+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTextChatModal/components/TextChatBubbleMe/TextChatBubbleMe.styles.tsx
