+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTextChatModal/components/TextChatBubbleMe/TextChatBubbleMe.tsx
@@ -1,4 +1,4 @@
-import { ChatRecivMessage } from "../AlgorithmTextChatModal/AlgorithmTextChatModal.types"
+import { ChatRecivMessage } from "../../AlgorithmTextChatModal.types"
 import * as S from "./TextChatBubbleMe.styles"
 
 export const TextChatBubbleMe = ({message}: {message: ChatRecivMessage}) => {
