+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTextChatModal/components/TextChatBubbleOthers/TextChatBubbleOthers.styles.tsx
