+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTextChatModal/components/TextChatBubbleOthers/TextChatBubbleOthers.tsx
@@ -1,4 +1,4 @@
-import { ChatRecivMessage } from "../AlgorithmTextChatModal/AlgorithmTextChatModal.types"
+import { ChatRecivMessage } from "../../AlgorithmTextChatModal.types"
 import * as S from "./TextChatBubbleOthers.styles"
 
 export const TextChatBubbleOthers = ({message}: {message: ChatRecivMessage}) => {
