+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmTextChatModal/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./TextChatBubbleMe/TextChatBubbleMe";
+export * from "./TextChatBubbleOthers/TextChatBubbleOthers";
