+++ idearly/src/pages/AlgorithmSolvingPage/components/index.ts
@@ -1,6 +1,3 @@
 export * from "./AlgorithmNav/AlgorithmNav";
 export * from "./AlgorithmVoiceChat/AlgorithmVoiceChat";
-export * from "./AlgorithmTextChat/AlgorithmTextChat";
 export * from "./AlgorithmTextChatModal/AlgorithmTextChatModal";
-export * from "./TextChatBubbleMe/TextChatBubbleMe";
-export * from "./TextChatBubbleOthers/TextChatBubbleOthers";
